rename test.sh -> do_raw_cal.sh. Update things. Do full FFT to get better sensitivity on weak satellites

+++ do_raw_cal.sh
@@ -13,17 +13,23 @@ TART_API="https://tart.elec.ac.nz/${TARGET}"
 : "${TART_CAL_POINTING:=0}"
 : "${TART_CAL_POINTING_RANGE:=3}"
 : "${TART_CAL_ARGS:=""}"
+: "${TART_GET_DATA:=1}"
+: "${TART_UPLOAD:=0}"
 
 DATA_DIR=./work_${TARGET}
 OUTPUT_DIR=./out_${TARGET} 
 
 # 
+if [ ${TART_GET_DATA} == 1 ]; then
+    echo "Downloading Data from ${TARGET}"
 
-rm -rf ${OUTPUT_DIR}
-mkdir -p ${OUTPUT_DIR}
-mkdir -p ${DATA_DIR}
-rm -rf ${DATA_DIR}/*
-python3 raw_cal/get_cal_data.py --api ${TART_API} --pw ${TART_LOGIN_PW} --n ${TART_NCAL} --interval ${TART_CAL_INT} --dir ${DATA_DIR}
+    rm -rf ${OUTPUT_DIR}
+    mkdir -p ${OUTPUT_DIR}
+    mkdir -p ${DATA_DIR}
+    rm -rf ${DATA_DIR}/*
+    python3 raw_cal/get_cal_data.py --api ${TART_API} --pw ${TART_LOGIN_PW} --n ${TART_NCAL} --interval ${TART_CAL_INT} --dir ${DATA_DIR}
+fi
+
 
 # Perform optimization
 # python3 raw_cal/pos_from_gps.py --api ${TART_API}  --elevation ${TART_CAL_ELEVATION}  --data ${DATA_DIR}  --dir ${OUTPUT_DIR}
@@ -39,7 +45,9 @@ python3 raw_cal/tart_cal.py --api ${TART_API}  \
 CALIB_OUTPUT=${OUTPUT_DIR}/BH_opt_json.json
 echo "Calibration output is in ${CALIB_OUTPUT}"
 
-tart_upload_gains --api ${TART_API} --gains ${CALIB_OUTPUT} --pw ${TART_LOGIN_PW}
-echo "Gains Uploaded"
-tart_upload_antenna_positions --api ${TART_API} --file ${CALIB_OUTPUT} --pw ${TART_LOGIN_PW}
-echo "Positions Uploaded"
+if [ ${TART_UPLOAD} == 1 ]; then
+    tart_upload_gains --api ${TART_API} --gains ${CALIB_OUTPUT} --pw ${TART_LOGIN_PW}
+    echo "Gains Uploaded"
+    tart_upload_antenna_positions --api ${TART_API} --file ${CALIB_OUTPUT} --pw ${TART_LOGIN_PW}
+    echo "Positions Uploaded"
+fi
